+++ packages/core/avatar/src/utils.js
@@ -1,14 +1,16 @@
 // @flow
-import type { ComponentType } from './types';
+import { type ComponentType } from 'react';
 
 export function omit(obj: {}, ...keysToOmit: Array<string>) {
-  return Object.keys(obj).reduce((acc, key: string) => {
-    if (keysToOmit.indexOf(key) === -1) acc[key] = obj[key];
-    return acc;
-  }, {});
+  const newObj = { ...obj };
+
+  for (const key of keysToOmit) {
+    delete newObj[key];
+  }
+  return newObj;
 }
 
-export function getDisplayName(prefix: string, Component: ComponentType) {
+export function getDisplayName(prefix: string, Component: ComponentType<*>) {
   const componentName: string = Component.displayName || Component.name;
 
   return componentName ? `${prefix}(${componentName})` : prefix;