+++ packages/core/datetime-picker/src/components/DatePicker.js
@@ -112,6 +112,29 @@ const StyledMenu = styled.div`
   z-index: ${layers.dialog};
 `;
 
+const Menu = ({ innerProps: menuInnerProps, selectProps }: Object) => (
+  <StyledMenu>
+    <Calendar
+      {...isoToObj(selectProps.calendarValue)}
+      {...isoToObj(selectProps.calendarView)}
+      disabled={selectProps.calendarDisabled}
+      onChange={selectProps.onCalendarChange}
+      onSelect={selectProps.onCalendarSelect}
+      // $FlowFixMe - Calendar is not a react component
+      ref={selectProps.calendarRef}
+      selected={[selectProps.selectedCalendarValue]}
+      innerProps={menuInnerProps}
+    />
+  </StyledMenu>
+);
+
+const FixedLayeredMenu = ({ selectProps, ...props }: Object) => (
+  <FixedLayer
+    containerRef={selectProps.calendarContainerRef}
+    content={<Menu {...props} selectProps={selectProps} />}
+  />
+);
+
 class DatePicker extends Component<Props, State> {
   // $FlowFixMe - Calendar isn't being correctly detected as a react component
   calendar: ElementRef<Calendar>;
@@ -154,6 +177,13 @@ class DatePicker extends Component<Props, State> {
       ...this.state,
       ...pick(this.props, ['value', 'isOpen']),
     };
+  };
+
+  /**
+   * Checks if a date is one of the disabled
+   */
+  isDateDisabled = (date: String) => {
+    return this.props.disabled.indexOf(date) > -1;
   };
 
   onCalendarChange = ({ iso }: { iso: string }) => {
@@ -214,7 +244,12 @@ class DatePicker extends Component<Props, State> {
     } else if (key === 'Backspace') {
       this.setState({ selectedValue: '' });
       this.triggerChange('');
-    } else if (key === 'Enter' || key === 'Tab') {
+
+      // Dates may be disabled
+    } else if (
+      !this.isDateDisabled(view) &&
+      (key === 'Enter' || key === 'Tab')
+    ) {
       this.triggerChange(view);
       this.setState({ isOpen: false, selectedValue: view });
     }
@@ -268,28 +303,19 @@ class DatePicker extends Component<Props, State> {
       this.props.appearance === 'subtle' || this.props.hideIcon
         ? null
         : this.props.icon;
-    const Menu = ({ innerProps: menuInnerProps }) => (
-      <StyledMenu>
-        <Calendar
-          {...isoToObj(value)}
-          {...isoToObj(view)}
-          disabled={disabled}
-          onChange={this.onCalendarChange}
-          onSelect={this.onCalendarSelect}
-          // $FlowFixMe - Calendar is not a react component
-          ref={this.refCalendar}
-          selected={[this.state.selectedValue]}
-          innerProps={menuInnerProps}
-        />
-      </StyledMenu>
-    );
 
-    const FixedLayeredMenu = props => (
-      <FixedLayer
-        containerRef={this.containerRef}
-        content={<Menu {...props} />}
-      />
-    );
+    const calendarProps = {
+      calendarContainerRef: this.containerRef,
+      calendarRef: this.refCalendar,
+      calendarDisabled: disabled,
+      calendarValue: value,
+      calendarView: view,
+      dropdownIndicatorIcon: icon,
+      onCalendarChange: this.onCalendarChange,
+      onCalendarSelect: this.onCalendarSelect,
+      selectedCalendarValue: this.state.selectedValue,
+    };
+
     const { styles: selectStyles = {} } = selectProps;
     const controlStyles =
       this.props.appearance === 'subtle' ? this.getSubtleControlStyles() : {};
@@ -316,7 +342,7 @@ class DatePicker extends Component<Props, State> {
           onFocus={this.onSelectFocus}
           components={{
             ClearIndicator,
-            DropdownIndicator: () => <DropdownIndicator icon={icon} />,
+            DropdownIndicator,
             Menu: FixedLayeredMenu,
           }}
           styles={mergeStyles(selectStyles, {
@@ -334,6 +360,7 @@ class DatePicker extends Component<Props, State> {
             }
           }
           {...selectProps}
+          {...calendarProps}
           spacing={spacing}
           validationState={validationState}
         />