+++ packages/core/datetime-picker/src/components/TimePicker.js
@@ -105,6 +105,15 @@ const menuStyles = {
   overflowY: 'auto',
 };
 
+const FixedLayerMenu = ({ selectProps, ...props }: Object) => {
+  return (
+    <FixedLayer
+      containerRef={selectProps.fixedLayerRef}
+      content={<components.Menu {...props} scrollMenuIntoView={false} />}
+    />
+  );
+};
+
 class TimePicker extends Component<Props, State> {
   containerRef: ?HTMLElement;
 
@@ -229,14 +238,6 @@ class TimePicker extends Component<Props, State> {
       this.props.appearance === 'subtle' || this.props.hideIcon
         ? null
         : this.props.icon;
-    const FixedLayerMenu = props => {
-      return (
-        <FixedLayer
-          containerRef={this.containerRef}
-          content={<components.Menu {...props} scrollMenuIntoView={false} />}
-        />
-      );
-    };
 
     const { styles: selectStyles = {}, ...otherSelectProps } = selectProps;
     const controlStyles =
@@ -255,7 +256,7 @@ class TimePicker extends Component<Props, State> {
           autoFocus={autoFocus}
           components={{
             ClearIndicator,
-            DropdownIndicator: () => <DropdownIndicator icon={icon} />,
+            DropdownIndicator,
             Menu: FixedLayerMenu,
           }}
           instanceId={id}
@@ -295,9 +296,11 @@ class TimePicker extends Component<Props, State> {
               value,
             }
           }
-          {...otherSelectProps}
           spacing={spacing}
+          dropdownIndicatorIcon={icon}
+          fixedLayerRef={this.containerRef}
           validationState={validationState}
+          {...otherSelectProps}
         />
       </div>
     );