+++ packages/editor/editor-common/src/ui/Extension/index.tsx
@@ -1,15 +1,13 @@
 import {
   akEditorFullPageMaxWidth,
   akEditorWideLayoutWidth,
+  akEditorBreakoutPadding,
 } from '../../styles';
-
-/** Full page padding, TODO: unify the padding values */
-const FULL_PAGE_PADDING = 66;
 
 export const calcExtensionWidth = (layout, containerWidth: number) => {
   switch (layout) {
     case 'full-width':
-      const effectiveFullWidth = containerWidth - FULL_PAGE_PADDING;
+      const effectiveFullWidth = containerWidth - akEditorBreakoutPadding;
       return effectiveFullWidth <= akEditorFullPageMaxWidth
         ? '100%'
         : `${effectiveFullWidth}px`;