+++ packages/editor/editor-core/src/ui/Appearance/Comment.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
-import styled, { keyframes } from 'styled-components';
+import styled from 'styled-components';
 import Button, { ButtonGroup } from '@atlaskit/button';
 import {
-  akColorR100,
   akColorN40,
   akBorderRadius,
   akGridSize,
@@ -16,24 +15,7 @@ import { pluginKey as maxContentSizePluginKey } from '../../plugins/max-content-
 import { stateKey as mediaPluginKey } from '../../plugins/media/pm-plugins/main';
 import { ClickAreaBlock } from '../Addon';
 import { tableCommentEditorStyles } from '../../plugins/table/ui/styles';
-
-const pulseBackground = keyframes`
-  50% {
-    background-color: ${akColorR100};
-  }
-`;
-
-const pulseBackgroundReverse = keyframes`
-  0% {
-    background-color: ${akColorR100};
-  }
-  50% {
-    background-color: auto;
-  }
-  100% {
-    background-color: ${akColorR100};
-  }
-`;
+import WithFlash from '../WithFlash';
 
 export interface CommentEditorProps {
   isMaxContentSizeReached?: boolean;
@@ -60,19 +42,10 @@ const CommentEditor: any = styled.div`
 
   max-width: inherit;
   word-wrap: break-word;
-
-  animation: ${(props: any) =>
-    props.isMaxContentSizeReached
-      ? `.25s ease-in-out ${pulseBackground}`
-      : 'none'};
-
-  &.-flash {
-    animation: 0.25s ease-in-out ${pulseBackgroundReverse};
-  }
 `;
 CommentEditor.displayName = 'CommentEditor';
 
-const TableControlsPadding = 24;
+const TableControlsPadding = 16;
 
 // tslint:disable-next-line:variable-name
 const MainToolbar = styled.div`
@@ -141,7 +114,6 @@ export default class Editor extends React.Component<
 > {
   static displayName = 'CommentEditorAppearance';
 
-  private flashToggle = false;
   private appearance: EditorAppearance = 'comment';
 
   private handleSave = () => {
@@ -178,15 +150,10 @@ export default class Editor extends React.Component<
     } = this.props;
     const maxContentSizeReached =
       maxContentSize && maxContentSize.maxContentSizeReached;
-    this.flashToggle = maxContentSizeReached && !this.flashToggle;
 
     return (
-      <div>
-        <CommentEditor
-          className={this.flashToggle ? '-flash' : ''}
-          isMaxContentSizeReached={maxContentSizeReached}
-          maxHeight={maxHeight}
-        >
+      <WithFlash animate={maxContentSizeReached}>
+        <CommentEditor maxHeight={maxHeight}>
           <MainToolbar>
             <Toolbar
               editorView={editorView!}
@@ -249,7 +216,7 @@ export default class Editor extends React.Component<
           <span style={{ flexGrow: 1 }} />
           {customSecondaryToolbarComponents}
         </SecondaryToolbar>
-      </div>
+      </WithFlash>
     );
   };