+++ packages/media/media-card/examples/card-files.tsx
@@ -9,7 +9,7 @@ import {
 } from '@atlaskit/media-test-helpers';
 import { AnalyticsListener } from '@atlaskit/analytics-next';
 
-import { UIAnalyticsEventInterface } from '../src/analytics-next';
+import { UIAnalyticsEventInterface } from '@atlaskit/analytics-next-types';
 
 import { Card, FileIdentifier } from '../src';
 import { createApiCards, actions } from '../example-helpers';