+++ packages/search/global-search/src/components/confluence/ConfluenceQuickSearchContainer.tsx
@@ -14,9 +14,11 @@ import renderSearchResults, {
   MAX_SPACES,
   MAX_PEOPLE,
 } from './ConfluenceSearchResults';
-import settlePromises from '../../util/settle-promises';
 import { LinkComponent } from '../GlobalQuickSearchWrapper';
-import { redirectToConfluenceAdvancedSearch } from '../SearchResultsUtil';
+import {
+  redirectToConfluenceAdvancedSearch,
+  handlePromiseError,
+} from '../SearchResultsUtil';
 import {
   ShownAnalyticsAttributes,
   buildShownEventDetails,
@@ -50,6 +52,7 @@ export interface State {
   objectResults: Result[];
   spaceResults: Result[];
   peopleResults: Result[];
+  keepRecentActivityResults: boolean;
 }
 
 /**
@@ -73,22 +76,28 @@ export class ConfluenceQuickSearchContainer extends React.Component<
       objectResults: [],
       spaceResults: [],
       peopleResults: [],
+      keepRecentActivityResults: true,
     };
   }
 
   handleSearch = (query: string) => {
-    this.setState({
-      query: query,
-    });
+    if (this.state.query !== query) {
+      this.setState({
+        query: query,
+        isLoading: true,
+      });
+    }
 
     if (query.length === 0) {
       // reset search results so that internal state between query and results stays consistent
       this.setState(
         {
           isError: false,
+          isLoading: false,
           objectResults: [],
           spaceResults: [],
           peopleResults: [],
+          keepRecentActivityResults: true,
         },
         () => this.fireShownPreQueryEvent(),
       );
@@ -107,13 +116,6 @@ export class ConfluenceQuickSearchContainer extends React.Component<
       query,
       this.state.searchSessionId,
     );
-
-    if (this.state.query === query) {
-      this.setState({
-        objectResults: results,
-      });
-    }
-
     return results;
   }
 
@@ -129,44 +131,22 @@ export class ConfluenceQuickSearchContainer extends React.Component<
         ----------------------------------
         Scope.ConfluencePageBlogAttachment,
         */
-
         Scope.ConfluenceSpace,
       ],
     );
-
-    if (this.state.query === query) {
-      this.setState({
-        /* 
-        TEMPORARILY DISABLED: XPSRCH-861 
-        ----------------------------------
-        // objectResults: results.get(Scope.ConfluencePageBlogAttachment) || [],
-        */
-
-        spaceResults: results.get(Scope.ConfluenceSpace) || [],
-      });
-    }
-
     return results;
   }
 
   async searchPeople(query: string): Promise<Result[]> {
     const results = await this.props.peopleSearchClient.search(query);
-
-    if (this.state.query === query) {
-      this.setState({
-        peopleResults: results,
-      });
-    }
-
     return results;
   }
 
   // TODO extract
-  handleSearchErrorAnalytics(source: string) {
-    return error => {
-      const { firePrivateAnalyticsEvent } = this.props;
-
-      if (firePrivateAnalyticsEvent) {
+  handleSearchErrorAnalytics(error, source: string): void {
+    const { firePrivateAnalyticsEvent } = this.props;
+    if (firePrivateAnalyticsEvent) {
+      try {
         firePrivateAnalyticsEvent(
           'atlassian.fabric.global-search.search-error',
           {
@@ -175,9 +155,16 @@ export class ConfluenceQuickSearchContainer extends React.Component<
             source: source,
           },
         );
+      } catch (error) {
+        // TODO logging on error
       }
-    };
+    }
   }
+
+  handleSearchErrorAnalyticsThunk = (
+    source: string,
+  ): ((reason: any) => void) => error =>
+    this.handleSearchErrorAnalytics(error, source);
 
   fireShownPreQueryEvent(requestStartTime?: number) {
     const { createAnalyticsEvent } = this.props;
@@ -218,72 +205,89 @@ export class ConfluenceQuickSearchContainer extends React.Component<
       );
     }
   }
+  getSearchResultState = (
+    query: string,
+    objectResults: Result[],
+    spaceResults: Map<Scope, Result[]>,
+    peopleResults: Result[],
+  ) => {
+    if (this.state.query === query) {
+      return {
+        objectResults,
+        spaceResults: spaceResults.get(Scope.ConfluenceSpace) || [],
+        peopleResults,
+      };
+    }
+    return {
+      objectResults: this.state.objectResults,
+      spaceResults: this.state.spaceResults,
+      peopleResults: this.state.peopleResults,
+    };
+  };
 
   doSearch = async (query: string) => {
     const startTime = performanceNow();
-    const quickNavPromise = this.searchQuickNav(query);
-    const confXpSearchPromise = this.searchCrossProductConfluence(query);
-    const searchPeoplePromise = this.searchPeople(query);
 
-    // trigger error analytics when a search fails
-    quickNavPromise.catch(
-      this.handleSearchErrorAnalytics('confluence.quicknav'),
+    this.setState({
+      isLoading: true,
+    });
+    const quickNavPromise = this.searchQuickNav(query).catch(error => {
+      this.handleSearchErrorAnalytics(error, 'confluence.quicknav');
+      // rethrow to fail the promise
+      throw error;
+    });
+    const confXpSearchPromise = handlePromiseError(
+      this.searchCrossProductConfluence(query),
+      new Map<Scope, Result[]>(),
+      this.handleSearchErrorAnalyticsThunk('xpsearch-confluence'),
     );
-    confXpSearchPromise.catch(
-      this.handleSearchErrorAnalytics('xpsearch-confluence'),
+
+    const searchPeoplePromise = handlePromiseError(
+      this.searchPeople(query),
+      [],
+      this.handleSearchErrorAnalyticsThunk('search-people'),
     );
-    searchPeoplePromise.catch(this.handleSearchErrorAnalytics('search-people'));
 
-    /*
-    * Handle error state
-    */
-    (async () => {
-      try {
-        await quickNavPromise;
-        this.setState({
-          isError: false,
-        });
-      } catch (e) {
-        this.setState({
-          isError: true,
-        });
-      }
-    })();
+    try {
+      const [
+        objectResults,
+        spaceResultsMap = new Map<Scope, Result[]>(),
+        peopleResults = [],
+      ] = await Promise.all([
+        quickNavPromise,
+        confXpSearchPromise,
+        searchPeoplePromise,
+      ]);
 
-    // handle loading state. true at the beginning, false only after all promises have settled.
-    (async () => {
-      try {
-        this.setState({
-          isLoading: true,
-        });
+      const searchResult = this.getSearchResultState(
+        query,
+        objectResults,
+        spaceResultsMap,
+        peopleResults,
+      );
 
-        const [
-          objectResults,
-          spaceResultsMap,
-          peopleResults,
-        ] = await settlePromises([
-          quickNavPromise,
-          confXpSearchPromise,
-          searchPeoplePromise,
-        ]);
+      this.setState({
+        isError: false,
+        isLoading: false,
+        keepRecentActivityResults: false,
+        ...searchResult,
+      });
 
-        // pull the space results out of the Map
-        const spaceResults = spaceResultsMap.get(Scope.ConfluenceSpace) || [];
-
-        this.fireShownPostQueryEvent(
-          startTime,
-          buildShownEventDetails(
-            take(objectResults, MAX_PAGES_BLOGS_ATTACHMENTS),
-            take(spaceResults, MAX_SPACES),
-            take(peopleResults, MAX_PEOPLE),
-          ),
-        );
-      } finally {
-        this.setState({
-          isLoading: false,
-        });
-      }
-    })();
+      this.fireShownPostQueryEvent(
+        startTime,
+        buildShownEventDetails(
+          take(searchResult.objectResults, MAX_PAGES_BLOGS_ATTACHMENTS),
+          take(searchResult.spaceResults, MAX_SPACES),
+          take(searchResult.peopleResults, MAX_PEOPLE),
+        ),
+      );
+    } catch {
+      this.setState({
+        isError: true,
+        isLoading: false,
+        keepRecentActivityResults: false,
+      });
+    }
   };
 
   handleMount = async () => {
@@ -293,32 +297,40 @@ export class ConfluenceQuickSearchContainer extends React.Component<
       isLoading: true,
     });
 
-    const recentItemsPromise = this.props.confluenceClient.getRecentItems(
-      this.state.searchSessionId,
-    );
-    const recentSpacesPromise = this.props.confluenceClient.getRecentSpaces(
-      this.state.searchSessionId,
-    );
-    const recentPeoplePromise = this.props.peopleSearchClient.getRecentPeople();
+    const sessionId = this.state.searchSessionId;
+    const { confluenceClient, peopleSearchClient } = this.props;
 
-    recentItemsPromise.catch(
-      this.handleSearchErrorAnalytics('recent-confluence-items'),
+    const recentActivityPromisesMap = {
+      'recent-confluence-items': confluenceClient.getRecentItems(sessionId),
+      'recent-confluence-spaces': confluenceClient.getRecentSpaces(sessionId),
+      'recent-people': peopleSearchClient.getRecentPeople(),
+    };
+
+    const recentActivityPromises = Object.keys(recentActivityPromisesMap).map(
+      key =>
+        handlePromiseError(
+          recentActivityPromisesMap[key],
+          [],
+          this.handleSearchErrorAnalyticsThunk(key),
+        ),
     );
-    recentSpacesPromise.catch(
-      this.handleSearchErrorAnalytics('recent-confluence-spaces'),
-    );
-    recentPeoplePromise.catch(this.handleSearchErrorAnalytics('recent-people'));
 
     try {
+      const [
+        recentlyViewedPages = [],
+        recentlyViewedSpaces = [],
+        recentlyInteractedPeople = [],
+      ] = await Promise.all(recentActivityPromises);
       this.setState(
         {
-          recentlyViewedPages: await recentItemsPromise,
-          recentlyViewedSpaces: await recentSpacesPromise,
-          recentlyInteractedPeople: await recentPeoplePromise,
+          recentlyViewedPages,
+          recentlyViewedSpaces,
+          recentlyInteractedPeople,
+          isLoading: false,
         },
         () => this.fireShownPreQueryEvent(startTime),
       );
-    } finally {
+    } catch {
       this.setState({
         isLoading: false,
       });
@@ -334,14 +346,15 @@ export class ConfluenceQuickSearchContainer extends React.Component<
     const {
       query,
       isLoading,
+      searchSessionId,
       isError,
-      recentlyViewedPages,
-      recentlyViewedSpaces,
-      recentlyInteractedPeople,
       objectResults,
       spaceResults,
       peopleResults,
-      searchSessionId,
+      recentlyViewedPages,
+      recentlyViewedSpaces,
+      recentlyInteractedPeople,
+      keepRecentActivityResults,
     } = this.state;
 
     return (
@@ -355,16 +368,17 @@ export class ConfluenceQuickSearchContainer extends React.Component<
         searchSessionId={searchSessionId}
       >
         {renderSearchResults({
+          retrySearch: this.retrySearch,
           query,
           isError,
-          isLoading,
-          retrySearch: this.retrySearch,
-          recentlyViewedPages,
-          recentlyViewedSpaces,
-          recentlyInteractedPeople,
           objectResults,
           spaceResults,
           peopleResults,
+          isLoading,
+          recentlyViewedPages,
+          recentlyViewedSpaces,
+          recentlyInteractedPeople,
+          keepRecentActivityResults,
           searchSessionId,
         })}
       </GlobalQuickSearch>